Update Trimmer.java
update aspectRatio

+++ android/src/main/java/com/shahenlibrary/Trimmer/Trimmer.java
@@ -261,8 +261,8 @@ public class Trimmer {
       int height = Integer.parseInt(retriever.extractMetadata(FFmpegMediaMetadataRetriever.METADATA_KEY_VIDEO_HEIGHT));
       int orientation = Integer.parseInt(retriever.extractMetadata(FFmpegMediaMetadataRetriever.METADATA_KEY_VIDEO_ROTATION));
 
-      float aspectRatio = width / height;
-
+      float aspectRatio = (float)width / (float)height;
+      
       int resizeWidth = 200;
       int resizeHeight = Math.round(resizeWidth / aspectRatio);
